requirements.txt
kafka-python
faker
faker-commerce
shortuuid